FROMLIST: HID: multitouch: report MT_TOOL_PALM for non-confident touches
According to Microsoft specification [1] for Precision Touchpads (and
Touchscreens) the devices use "confidence" reports to signal accidental
touches, or contacts that are "too large to be a finger". Instead of
simply marking contact inactive in this case (which causes issues if
contact was originally proper and we lost confidence in it later, as
this results in accidental clicks, drags, etc), let's report such
contacts as MT_TOOL_PALM and let userspace decide what to do.
Additionally, let's report contact size for such touches as maximum
allowed for major/minor, which should help userspace that is not yet
aware of MT_TOOL_PALM to still perform palm rejection.

An additional complication, is that some firmwares do not report
non-confident touches as active. To cope with this we delay release of
such contact (i.e. if contact was active we first report it as still
active MT+TOOL_PALM and then synthesize the release event in a separate
frame).

[1] https://docs.microsoft.com/en-us/windows-hardware/design/component-guidelines/touchpad-windows-precision-touchpad-collection

Signed-off-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>

BUG=b:38045083
TEST=Build and boot Eve, try using touchscreen/touchpad

(am from https://patchwork.kernel.org/patch/9894803/)
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Change-Id: I99755f0f5a3673c52cdb5293d00f168345aa7a26
Reviewed-on: https://chromium-review.googlesource.com/617811
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>

[rebase414(lannm): Fixed conflicts]
Signed-off-by: Lann Martin <lannm@chromium.org>