UPSTREAM: clockevents: Update clockevents device next_event on stop
clockevent_device::next_event holds the next timer event of a clock event
device. The value is updated in clockevents_program_event(), i.e. when the
hardware timer is armed for the next expiry.

When there are no software timers armed on a CPU, the corresponding per CPU
clockevent device is brought into ONESHOT_STOPPED state, but
clockevent_device::next_event is not updated, because
clockevents_program_event() is not called.

So the content of clockevent_device::next_event is stale, which is not an
issue when real hardware is used. But the hrtimer broadcast device relies
on that information and the stale value causes spurious wakeups.

Update clockevent_device::next_event to KTIME_MAX when it has been brought
into ONESHOT_STOPPED state to avoid spurious wakeups. This reflects the
proper expiry time of the stopped timer: infinity.

[ tglx: Massaged changelog ]

Signed-off-by: Prasad Sodagudi <psodagud@codeaurora.org>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Cc: viresh.kumar@linaro.org
Link: https://lkml.kernel.org/r/1509043042-32486-1-git-send-email-psodagud@codeaurora.org

BUG=chromium:795946
TEST=Build and boot

Change-Id: I3bb1d17bd11df30f1dfcf99d1864e2720c42e9c1
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit 39c82caff8610d57ffe32157cb3130dfabe12fbe)
Reviewed-on: https://chromium-review.googlesource.com/843480
Tested-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>