UPSTREAM: tpm: use struct tpm_chip for tpm_chip_find_get()
Device number (the character device index) is not a stable identifier
for a TPM chip. That is the reason why every call site passes
TPM_ANY_NUM to tpm_chip_find_get().

This commit changes the API in a way that instead a struct tpm_chip
instance is given and NULL means the default chip. In addition, this
commit refines the documentation to be up to date with the
implementation.

Suggested-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com> (@chip_num -> @chip part)
Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@linux.intel.com>
Reviewed-by: Jason Gunthorpe <jgg@ziepe.ca>
Tested-by: PrasannaKumar Muralidharan <prasannatsmkumar@gmail.com>

BUG=chromium:822509
TEST=Build kernel for Cheza, which uses mocktpm

Signed-off-by: Evan Green <evgreen@chromium.org>
(cherry picked from commit aad887f6641145fec2a801da2ce4ed36cf99c6a5)
Change-Id: I3f9ea1545f0023107ad70aaeed58b3562c9d5c2a
Reviewed-on: https://chromium-review.googlesource.com/966873
Reviewed-by: Guenter Roeck <groeck@chromium.org>