UPSTREAM: ASoC: rt5663: Delay and retry reading rt5663 ID register
In the probe, the codec may not be ready for I2C reading or there are some
glitches on the i2c line. So if the i2c reading value is incorrect, it will
read again after delay. This issue is similar the patch
https://patchwork.kernel.org/patch/9681421/. In current project, these 2
devices were connected to the same i2c line, and they met the same problem.

Signed-off-by: Oder Chiou <oder_chiou@realtek.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit ba68fa318442d25586d057c1e8c9165bdc62c1b3)

BUG=b:80592327
TEST=ARCH=arm64 make allyesconfig;
     CROSS_COMPILE=aarch64-cros-linux-gnu- ARCH=arm64 make -j64

Change-Id: I006f9540adc450ba4e29033854d13d59686873da
Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1139717