UPSTREAM: drm/amdkfd: Add CWSR support
This hardware feature allows the GPU to preempt shader execution in
the middle of a compute wave, save the state and restore it later
to resume execution.

Memory for saving the state is allocated per queue in user mode and
the address and size passed to the create_queue ioctl. The size
depends on the number of waves that can be in flight simultaneously
on a given ASIC.

BUG=b:79158447
TEST=boot Grunt, run graphics_SanAngeles and login_LoginSuccess

Signed-off-by: Shaoyun.liu <shaoyun.liu@amd.com>
Signed-off-by: Yong Zhao <yong.zhao@amd.com>
Signed-off-by: Felix Kuehling <Felix.Kuehling@amd.com>
Signed-off-by: Oded Gabbay <oded.gabbay@gmail.com>
(cherry picked from commit 373d7080896a3cb3b28ae3a2abdafb7bb87552b1)
Signed-off-by: Dominik Behr <dbehr@chromium.org>

Change-Id: I3e762b579356a84ae4b5a2ff7a499b82d9da903c
Reviewed-on: https://chromium-review.googlesource.com/1137606
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>