CHROMIUM: iwl7000: mvm: avoid dumping assert log when device is stopped
We might erroneously get to error dumping code when the
device is already stopped.
In that case the driver will detect a defective value and
will try to reset the HW, assuming it is only a bus issue.
The driver than proceeds with the dumping.
The result has two side effects:
The device won't be stopped again, since the transport
status is already stopped, so the device remains powered
on while it actually should be stopped.
In addition, the dump in that case is completely garbaged
and useless.
Detect and avoid this.
It will also make debugging such issues easier.

Signed-off-by: Sara Sharon <sara.sharon@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: f5ac5e2f5f23b106e3ed654d67f51e45389df51a