UPSTREAM: clk: use round rate to bail out early in set_rate
The current implementation of clk_core_set_rate_nolock() bails out early
if the requested rate is exactly the same as the one set. It should bail
out if the request would not result in a rate a change. This is important
when the rate is not exactly what is requested, which is fairly common
with PLLs.

Ex: provider able to give any rate with steps of 100Hz
 - 1st consumer request 48000Hz and gets it.
 - 2nd consumer request 48010Hz as well. If we were to perform the usual
   mechanism, we would get 48000Hz as well. The clock would not change so
   there is no point performing any checks to make sure the clock can
   change, we know it won't.

This is important to prepare the addition of the clock protection
mechanism

Acked-by: Linus Walleij <linus.walleij@linaro.org>
Tested-by: Quentin Schulz <quentin.schulz@free-electrons.com>
Tested-by: Maxime Ripard <maxime.ripard@free-electrons.com>
Acked-by: Michael Turquette <mturquette@baylibre.com>
Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>
Signed-off-by: Michael Turquette <mturquette@baylibre.com>
Link: lkml.kernel.org/r/20171201215200.23523-6-jbrunet@baylibre.com
(cherry picked from commit ca5e089a32c5ffba6c5101fdabdd6dea18041c34)
Signed-off-by: Akshu Agrawal <akshu.agrawal@amd.com>

BUG=b:123433680
TEST=
Play audio on da7219 codec which is headphone and see no errors.
Check for 64 bclks per wclk.
cat /sys/kernel/debug/regmap/i2c-DLGS7219\:00/registers | grep 2b
2b: 81

Change-Id: I02c2591206acb27b130d2a12a7a82bcef4ff9f4c
Reviewed-on: https://chromium-review.googlesource.com/1587702
Commit-Ready: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Akshu Agrawal <akshu.agrawal@amd.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>