FROMLIST: pinctrl: cherryview: Mask all interrupts on Intel_Strago based systems
Guenter Roeck reported an interrupt storm on a prototype system which is
based on Cyan Chromebook. The root cause turned out to be a incorrectly
configured pin that triggers spurious interrupts. This will be fixed in
coreboot but currently we need to prevent the interrupt storm from
happening by masking all interrupts (but not GPEs) on those systems.

BUG=chromium:791604
TEST=Build and run image on Cyan

Change-Id: I6c43482dedc75a51c21daa9c4bef3e6e0ff5ccb5
Link: https://bugzilla.kernel.org/show_bug.cgi?id=197953
Fixes: bcb48cca23ec ("pinctrl: cherryview: Do not mask all interrupts in probe")
Reported-and-tested-by: Guenter Roeck <linux@roeck-us.net>
Reported-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Cc: stable@vger.kernel.org
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(am from https://patchwork.ozlabs.org/patch/844155/)
Reviewed-on: https://chromium-review.googlesource.com/791891
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>