UPSTREAM: drm/sched: move fence slab handling to module init/exit
This is the only part of the scheduler which must not be called from
different drivers. Move it to module init/exit so it is done a single
time when loading the scheduler.

BUG=b:79158447
TEST=boot Grunt, run graphics_SanAngeles and login_LoginSuccess

Reviewed-by: Chunming Zhou <david1.zhou@amd.com>
Tested-by: Dieter Ntzel <Dieter@nuetzel-hh.de>
Acked-by: Alex Deucher <alexander.deucher@amd.com>
Signed-off-by: Lucas Stach <l.stach@pengutronix.de>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
(cherry picked from commit 4983e48c8539282be15f660bdd2c4260467b1190)
Signed-off-by: Dominik Behr <dbehr@chromium.org>

Change-Id: I31c5cceb85cb318af8f1f8d7b0abdd441c0dc7ff
Reviewed-on: https://chromium-review.googlesource.com/1134233
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Dominik Behr <dbehr@google.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>