UPSTREAM: remoteproc: Move resource table load logic to find
Extend the previous operation of finding the resource table in the ELF
with the extra step of populating the rproc struct with a copy and the
size. This allows drivers to override the mechanism used for acquiring
the resource table, or omit it for firmware that is known not to have a
resource table.

This leaves the custom, dummy, find_rsc_table implementations found in
some drivers dangling.

Reviewed-By: Loic Pallardy <loic.pallardy@st.com>
Tested-By: Loic Pallardy <loic.pallardy@st.com>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>

BUG=chromium:795946
TEST=Build and boot

Change-Id: I81d1701e65ddd8272c9a3fd8db9a921e0ffe54fd
Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>
(cherry picked from commit 58b640906702bcc083ac783bf10325e22c67d9fc)
Reviewed-on: https://chromium-review.googlesource.com/1016192
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>