UPSTREAM: net: qualcomm: rmnet: Define the MAPv4 packet formats
The MAPv4 packet format adds support for RX / TX checksum offload.
For a bi-directional UDP stream at a rate of 570 / 146 Mbps, roughly
10% CPU cycles are saved.

For receive path, there is a checksum trailer appended to the end of
the MAP packet. The valid field indicates if hardware has computed
the checksum. csum_start_offset indicates the offset from the start
of the IP header from which hardware has computed checksum.
csum_length is the number of bytes over which the checksum was
computed and the resulting value is csum_value.

In the transmit path, a header is appended between the end of the MAP
header and the start of the IP packet. csum_start_offset is the offset
in bytes from which hardware will compute the checksum if the
csum_enabled bit is set. udp_ip4_ind indicates if the checksum
value of 0 is valid or not. csum_insert_offset is the offset from the
csum_start_offset where hardware will insert the computed checksum.

The use of this additional packet format for checksum offload is
explained in subsequent patches.

Signed-off-by: Subash Abhinov Kasiviswanathan <subashab@codeaurora.org>
Signed-off-by: David S. Miller <davem@davemloft.net>

BUG=chromium:795946
TEST=Build and boot

Change-Id: I070e6098dc59e40cde005032dea648b67bf99b4a
(cherry picked from commit c597897b08cb56e6a8b240245c074d43d1fde735)
Reviewed-on: https://chromium-review.googlesource.com/897922
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Subash Abhinov Kasiviswanathan <subashab@codeaurora.org>
Reviewed-by: Ben Chan <benchan@chromium.org>