UPSTREAM: drm/i915: Add function to output Aksv over GMBUS
Once the Aksv is available in the PCH, we need to get it on the wire to
the receiver via DDC. The hardware doesn't allow us to read the value
directly, so we need to tell GMBUS to source the Aksv internally and
send it to the right offset on the receiver.

The way we do this is to initiate an indexed write where the index is
the Aksv register offset. We write dummy values to GMBUS3 as if we were
sending the key, and the hardware slips in the "real" values when it
goes out.

Changes in v2:
- None
Changes in v3:
- Uses new index write feature (Ville)
Changes in v4:
- None
Changes in v5:
- checkpatch whitespace fix
Changes in v6:
- None

Cc: Ville Syrjälä <ville.syrjala@linux.intel.com>
Reviewed-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Link: https://patchwork.freedesktop.org/patch/msgid/20180108195545.218615-8-seanpaul@chromium.org

BUG=b:141971239
TEST=boot Meep, run login_LoginSuccess

Signed-off-by: Chiou, Cooper <cooper.chiou@intel.com>
(cherry picked from commit 07e17a75922a593393b7eb8f73622723b76d3706)

Change-Id: I56a1c7781448abeb905107445df521a042dad64c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1857842
Tested-by: Cooper Chiou <cooper.chiou@intel.com>
Tested-by: Henry Sun <henrysun@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Henry Sun <henrysun@google.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Auto-Submit: Cooper Chiou <cooper.chiou@intel.com>