UPSTREAM: clk: Print the clock name and warning cause
When getting the clock related warnings, it is useful to know what
is the clock name that is causing the problem and the cause of the
problem.

Add the clock name and the the warning cause to the log, so that the
output becomes clearer like this:

[    2.383969] ------------[ cut here ]------------
[    2.388720] WARNING: CPU: 0 PID: 1 at drivers/clk/clk.c:814 clk_core_disable+0xd4/0xf8
[    2.396658] uart4_ipg_gate already disabled

Signed-off-by: Fabio Estevam <fabio.estevam@nxp.com>
Reviewed-by: Geert Uytterhoeven <geert+renesas@glider.be>
Signed-off-by: Stephen Boyd <sboyd@kernel.org>
(cherry picked from commit ab525dcccfd6ad11a0313fd58222b72cee4ca7ad)
Signed-off-by: Akshu Agrawal <akshu.agrawal@amd.com>

BUG=b:123433680
TEST=
Play audio on da7219 codec which is headphone and see no errors.
Check for 64 bclks per wclk.
cat /sys/kernel/debug/regmap/i2c-DLGS7219\:00/registers | grep 2b
2b: 81

Change-Id: If931bc65a18d77148198965bf09ae6f190368f8a
Reviewed-on: https://chromium-review.googlesource.com/1587707
Commit-Ready: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Akshu Agrawal <akshu.agrawal@amd.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>