BACKPORT: drm/i915: Reinitialize sink scrambling/TMDS clock ratio on HPD
The LG 4k TV I have doesn't deassert HPD when I turn the TV off, but
when I turn it back on it will pulse the HPD line. By that time it has
forgotten everything we told it about scrambling and the clock ratio.
Hence if we want to get a picture out if it again we have to tell it
whether we're currently sending scrambled data or not. Implement
that via the encoder->hotplug() hook.

v2: Force a full modeset to not follow the HDMI 2.0 spec more
    closely (Shashank)

[pushed with whitespace fixes to make sparse happy]
Cc: Shashank Sharma <shashank.sharma@intel.com>
Cc: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Signed-off-by: Lyude Paul <lyude@redhat.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20180117192149.17760-1-ville.syrjala@linux.intel.com

(cherry picked from commit dba14b27dd3ca5ce5b3a1d538862e7dce556dba7)

Conflicts:
   drivers/gpu/drm/i915/intel_crt.c
   drivers/gpu/drm/i915/intel_ddi.c
   drivers/gpu/drm/i915/intel_drv.h

BUG=b:204950015
TEST=Connect external display through MST and both screens turn on

Signed-off-by: Mark Yacoub <markyacoub@google.com>
Change-Id: Idc21068d925dd70186c8e7513701c1ad49073c62
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3412704
Reviewed-by: Sean Paul <seanpaul@chromium.org>