UPSTREAM: drm/amdgpu: add optional ring to *_hdp callbacks
This adds an optional ring to the invalidate_hdp and flush_hdp
callbacks. If the ring isn't specified or the emit_wreg function not
available the HDP operation will be done with the CPU otherwise by
writing on the ring.

BUG=b:112542653
TEST=boot Grunt, run graphics_SanAngeles and login_LoginSuccess

Signed-off-by: Christian König <christian.koenig@amd.com>
Acked-by: Chunming Zhou <david1.zhou@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
(cherry picked from commit 698825653fdf1a696e1b9458ed9fc4aa2c6587d4)
Signed-off-by: Dominik Behr <dbehr@chromium.org>

Change-Id: I00616b9b66ffc1513ad3d1d729773f21980ce7cf