CHROMIUM: grunt: sdhci: Disable BH720/O2.
The BayhubTech 720/SEABIRD0 PCI device id 0x8620 is currently
claimed by the in-tree SDHCI driver.  However, it misbehaves and
spams the kernel log with frequent "Timeout waiting for hardware
cmd interrupt" messages.  These messages reduce the usability of
the log on devices with this chip and are generally annoying.

To reduce the log-spam in the kernel log and disassociate the driver
from the incompatible SDHCI controller, make the probe function
explicitly fail when this device is probed.

Ideally, we would just remove the pci_ids structure that lists
compatible devices and never load/probe this device.  However, the
call to PCI_DEVICE_CLASS matches this device causes this device to
be probed even with the removal of it's PCI vendor/device ID.

This commit should be reverted when b/71875355 is fixed.

BUG=b:72120626 b:71875355
TEST=Installed on grunt, verified that message disappears.

Signed-off-by: Craig Bergstrom <criagb@google.com>
Change-Id: I2bdfe06eb15cf3bb7080ed41d4d2fcc9b455f980
Reviewed-on: https://chromium-review.googlesource.com/879243
Commit-Ready: Craig Bergstrom <craigb@chromium.org>
Tested-by: Craig Bergstrom <craigb@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>