Revert "CHROMIUM: Revert "objtool: Allow alternatives to be ignored""
This reverts commit e9eb28165cd73ba2355d77cd14d9d63869123073.

Upstream stable release v4.14.18 fixes the problem we avoided by reverting
the original patch. The problem was only seen in chromeos-4.14 if retpoline
support was enabled through KCFLAGS. This is no longer the case (retpoline
support is now enabled with CONFIG_RETPOLINE), thus it is safe to restore
the original patch. This is a prerequisite for the merge of v4.14.18.

BUG=chromium:800667, chromium:802348, chromium:810185
TEST=Build and run on affected system running chromeos-4.14

Change-Id: I010644f9f84d50d24c733147a754919a1ed8515f
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/908532
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>