FROMGIT: amdgpu/dc/dml: Support clang option for stack alignment
DML uses the compiler option -mpreferred-stack-boundary=4 to configure
a stack alignment of 16 bytes. Clang uses the option -mstack-alignment
instead, which expects as parameter the alignment in bytes, and not a
power of two like -mpreferred-stack-boundary.

Probe for both compiler options and use the correct one, similar to
what is done in arch/x86/Makefile.

Reported-by: Guenter Roeck <groeck@chromium.org>
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Harry Wentland <harry.wentland@amd.com>
(cherry picked from
commit c73afced6a1550c2151b1f8de7effdaa997375f4)

BUG=chromium:809829 people.freedesktop.org ~agd5f/linux amd-staging-drm-next
TEST=build for kahlee with clang

Change-Id: Ic82082b5b7a3f26aa8126b747ee5f068a05217dd
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/910252
Reviewed-by: Guenter Roeck <groeck@chromium.org>