UPSTREAM: soc: qcom: rmtfs-mem: Add support for assigning memory to remote
On some platform the remote processor's memory map is not statically
configured in TrustZone, so each memory region that is to be accessed by
the remote needs a call into TrustZone to set up the remote's
permissions.

Implement this for the rmtfs memory driver, to give the modem on 8996
access to the shared file system buffers.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Andy Gross <andy.gross@linaro.org>
(cherry picked from commit fa65f8045137d061134c15233d35a089cc85b87e)

BUG=chromium:795946
BUG=b:79668211
TEST=Build and boot cheza.

Change-Id: Ie7876df1eb0edc096685359a302cc8ca9a79deb4
Signed-off-by: Ben Chan <benchan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1119465
Reviewed-by: Douglas Anderson <dianders@chromium.org>