UPSTREAM: ASoC: dmic: Add optional wakeup delay
On some systems a delay is needed after switching on the clocks, to allow
the output to stabilize and avoid a popping noise at the beginning of
the recording. Add the optional device tree property 'wakeup-delay-ms'
and apply the specified delay after enabling the mic. A blocking delay
can't be applied in dmic_daiops_trigger() since the function is called
in atomic context. Instead use a DAPM event handler to set the enable
GPIO and apply the delay in the handler.

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 05c9b302eda71083840392d74ce62dd1e1f30621)
Signed-off-by: Samaga Prasanna Krishna <samaga.krishna@intel.corp-partner.google.com>

BUG=b:110025789
TEST=Test Audio use cases with full SOF 0.6 rebase release series applied.

Change-Id: I4b0dd6a251b7d12ef62e833f4c42fbe3ca4d1522
Reviewed-on: https://chromium-review.googlesource.com/1353412
Commit-Ready: Curtis Malainey <cujomalainey@chromium.org>
Tested-by: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>