UPSTREAM: drm/msm/mdp5: rework CTL START signal handling
For DSI cmd-mode and writeback, we need to write the CTL's START
register to kick things off, but we only want to do that once both
the encoder and the crtc have a chance to write their corresponding
flush bits.  The difficulty is that when there is a full modeset
(ie. encoder state has changed) we want to defer the start until
encoder->enable().  But if only plane's have changed, we want to do
this from crtc->commit().

The start_mask was a previous attempt to handle this, but it didn't
really do the right thing since atomic conversion.

Instead track in the crtc state that the start should be deferred,
set to try from encoder's (or in future writeback's) atomic_check().
This way the state is part of the atomic state, and rollback can
work properly if an atomic test fails.

Signed-off-by: Rob Clark <robdclark@gmail.com>
(cherry picked from commit f9cb8d8d836e155f361c3f1bbe0802ae1f98a17e)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

BUG=b:80589817
TEST=Tested on cheza

Change-Id: I81584e8dea4b02bbda0fadf5d6fabd649e353e58
Reviewed-on: https://chromium-review.googlesource.com/1158886
Commit-Ready: Sean Paul <seanpaul@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Kristian H. Kristensen <hoegsberg@chromium.org>