CHROMIUM: Revert "Bluetooth: btusb: Restore QCA Rome suspend/resume fix with a "rewritten" version"
This reverts commit fed016a79b42943f7dea7409d107b641fe3d1080.

As per commit 629db59a0f14 ("CHROMIUM: Revert "UPSTREAM: Bluetooth: btusb:
driver to enable the usb-wakeup feature"") in chromeos-4.4, this patch results
in inband communication from the btusb device to cause the system to wake up.
While the hereby reverted patch is an attempt to fix that problem, it fails to
do so. Revert it prior to the merge of v4.14.20 into chromeos-4.14.

BUG=chromium:812853
TEST=Run system with Bluetooth enabled and verify wakeup functionality

Signed-off-by: Guenter Roeck <groeck@chromium.org>